external/safecoder_master/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1
transformers==4.32.0
accelerate
termcolor
yamlize
tabulate
libcst